fix(collection handler): Improved handling of movies without Tmdb ID's in Plex metadata.

+++ server/src/modules/collections/collection-worker.service.ts
@@ -138,40 +138,53 @@ export class CollectionWorkerService implements OnApplicationBootstrap {
 
     if (plexLibrary.type === 'movie') {
       if (this.settings.radarrConfigured()) {
-        const radarrMedia = await this.servarrApi.RadarrApi.getMovieByTmdbId(
-          media.tmdbId,
-        );
-        switch (collection.arrAction) {
-          case ServarrAction.DELETE:
-            await this.servarrApi.RadarrApi.deleteMovie(
-              radarrMedia.id,
-              true,
-              collection.listExclusions,
+        // find tmdbid
+        const tmdbid = media.tmdbId
+          ? media.tmdbId
+          : await this.tmdbIdService.getTmdbIdFromPlexRatingKey(
+              media.plexId.toString(),
             );
-            this.infoLogger('Removed movie from filesystem & Radarr');
-            break;
-          case ServarrAction.UNMONITOR:
-            await this.servarrApi.RadarrApi.unmonitorMovie(
-              radarrMedia.id,
-              false,
-            );
-            this.infoLogger('Unmonitored movie in Radarr');
-            break;
-          case ServarrAction.DELETE_UNMONITOR_ALL:
-            await this.servarrApi.RadarrApi.unmonitorMovie(
-              radarrMedia.id,
-              true,
-            );
-            this.infoLogger('Unmonitored movie in Radarr & removed files');
-            break;
-          case ServarrAction.DELETE_UNMONITOR_EXISTING:
-            await this.servarrApi.RadarrApi.deleteMovie(
-              radarrMedia.id,
-              true,
-              collection.listExclusions,
-            );
-            this.infoLogger('Removed movie from filesystem & Radarr');
-            break;
+
+        if (tmdbid) {
+          const radarrMedia = await this.servarrApi.RadarrApi.getMovieByTmdbId(
+            media.tmdbId,
+          );
+          switch (collection.arrAction) {
+            case ServarrAction.DELETE:
+              await this.servarrApi.RadarrApi.deleteMovie(
+                radarrMedia.id,
+                true,
+                collection.listExclusions,
+              );
+              this.infoLogger('Removed movie from filesystem & Radarr');
+              break;
+            case ServarrAction.UNMONITOR:
+              await this.servarrApi.RadarrApi.unmonitorMovie(
+                radarrMedia.id,
+                false,
+              );
+              this.infoLogger('Unmonitored movie in Radarr');
+              break;
+            case ServarrAction.DELETE_UNMONITOR_ALL:
+              await this.servarrApi.RadarrApi.unmonitorMovie(
+                radarrMedia.id,
+                true,
+              );
+              this.infoLogger('Unmonitored movie in Radarr & removed files');
+              break;
+            case ServarrAction.DELETE_UNMONITOR_EXISTING:
+              await this.servarrApi.RadarrApi.deleteMovie(
+                radarrMedia.id,
+                true,
+                collection.listExclusions,
+              );
+              this.infoLogger('Removed movie from filesystem & Radarr');
+              break;
+          }
+        } else {
+          this.infoLogger(
+            `Couldn't find correct tmdbid. No action taken for movie: ${media.plexData.title}`,
+          );
         }
       }
     } else {
